requirements.txt
Django
djangorestframework
channels
Pillow
django-allauth
django-cors-headers
djangorestframework-simplejwt
daphne
uvicorn